Factor out complex ingredient handlers into their own modules.
This commit factors out the handlers for:
* `data.browser_compatibility`
* `data.specifications`
* `prose.short_description`

into their own modules. It also moves the top-level
"index-handlers.js" to index-handlers/index.js, to sit alongside
the handler modules.
