--- scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers.js
@@ -1,192 +1,0 @@
-const { select } = require("hast-util-select");
-const filter = require("unist-util-filter");
-const toString = require("hast-util-to-string");
-const visit = require("unist-util-visit");
-
-const utils = require("./utils.js");
-
-/**
- * Functions to check for recipe ingredients in Kuma page sources.
- *
- * The key is the name of a recipe ingredient (e.g.,
- * `data.browser_compatbiility` or `prose.syntax`) and the value is a function
- * to process a tree and file for that.
- *
- * Handler functions must take three arguments: a hast tree, a VFile, and a
- * context object. The context object has two entries:
- *
- * - `ingredient` - the name of the ingredient
- * - `recipeName` - the name of the recipe
- *
- * Handler functions may log messages against the file.
- *
- */
-const ingredientHandlers = {
-  default: (tree, file, context) => {
-    const { recipeName, ingredient, source } = context;
-    const rule = `${recipeName}/${ingredient}`;
-    const origin = `${source}:${rule}`;
-
-    file.message(`Linting ${ingredient} ingredient is unimplemented`, origin);
-  },
-  "data.browser_compatibility": (tree, file, context) => {
-    const id = "Browser_compatibility";
-    const body = select(`body`, tree);
-    const heading = select(`h2#${id}`, body);
-
-    if (heading === null) {
-      const message = file.message(
-        `Expected h2#${id}`,
-        body,
-        `${context.source}:${context.recipeName}/${context.ingredient}/expected-heading`
-      );
-      message.fatal = true;
-      utils.logMissingIngredient(file, context);
-      return;
-    }
-
-    let macroCount = 0;
-    visit(
-      utils.sliceSection(heading, body),
-      node => utils.isMacro(node, "Compat"),
-      () => {
-        macroCount += 1;
-      }
-    );
-
-    if (macroCount !== 1) {
-      utils.logMissingIngredient(file, context);
-    }
-  },
-  "data.examples": requireTopLevelHeading("Examples"),
-  "data.specifications": (tree, file, context) => {
-    const id = "Specifications";
-    const body = select(`body`, tree);
-
-    const heading = select(`h2#${id}`, body);
-    if (heading === null) {
-      const message = file.message(
-        `Expected h2#${id}`,
-        body,
-        `${context.source}:${context.recipeName}/${context.ingredient}/expected-heading`
-      );
-      message.fatal = true;
-      utils.logMissingIngredient(file, context);
-      return;
-    }
-
-    let sectionOk = false;
-    visit(utils.sliceSection(heading, body), "text", node => {
-      if (utils.isMacro(node, "SpecName")) {
-        sectionOk = true;
-        return visit.SKIP;
-      }
-
-      if (node.value.includes("Not part of any standard")) {
-        sectionOk = true;
-        return visit.SKIP;
-      }
-    });
-
-    if (!sectionOk) {
-      const message = file.message(
-        `Expected SpecName macro for ${context.recipeName}: ${context.ingredient}`,
-        heading,
-        `${context.source}:${context.recipeName}/${context.ingredient}/expected-macro`
-      );
-      message.fatal = true;
-      utils.logMissingIngredient(file, context);
-    }
-  },
-  "prose.description": requireTopLevelHeading("Description"),
-  "prose.error_type": requireTopLevelHeading("Error_type"),
-  "prose.message": requireTopLevelHeading("Message"),
-  "prose.see_also": requireTopLevelHeading("See_also"),
-  "prose.short_description": (tree, file, context) => {
-    // Short descriptions are complicated!
-    //
-    // A short description is understood to be either an seoSummary <span> or
-    // the first <p> that precedes:
-    //
-    // - an interactive example macro
-    // - an <h2>
-    // - the end of the document
-    //
-    // whichever comes first, but excluding <p>'s that are admonitions (warnings
-    // or notes).
-
-    const body = select("body", tree);
-
-    if (select("span.seoSummary", tree) !== null) {
-      return;
-    }
-
-    // Slice the tree to the nodes between the first element in <body> and
-    // a terminating node (interactive example or h2) or the end of the
-    // document
-    const introSection = utils.sliceBetween(
-      select(":first-child", body),
-      node => {
-        if (node.tagName === "h2") {
-          return true;
-        }
-
-        let containsInteractiveExample = false;
-        visit(
-          node,
-          node => utils.isMacro(node, "EmbedInteractiveExample"),
-          () => {
-            containsInteractiveExample = true;
-            return visit.EXIT;
-          }
-        );
-        return containsInteractiveExample;
-      },
-      body
-    );
-
-    // Remove admonition paragraphs
-    const isAdmonition = node =>
-      node.tagName === "p" &&
-      node.properties.className &&
-      (node.properties.className.includes("warning") ||
-        node.properties.className.includes("note"));
-    const filtered = filter(introSection, node => !isAdmonition(node));
-
-    // Get the first paragraph left over
-    const shortDescriptionP = select("p", filtered);
-
-    if (shortDescriptionP === null) {
-      utils.logMissingIngredient(file, context);
-      return;
-    }
-
-    // Check if the paragraph actually contains text
-    const shortDescriptionText = toString(shortDescriptionP).trim();
-
-    // See if there's any text remaining
-    if (!shortDescriptionText.length) {
-      utils.logMissingIngredient(file, context);
-    }
-  },
-  "prose.syntax": requireTopLevelHeading("Syntax"),
-  "prose.what_went_wrong": requireTopLevelHeading("What_went_wrong")
-};
-
-/**
- * A convenience function that returns ingredient handlers for checking the existence of a certain H2 in a hast tree.
- *
- * @param {String} ingredient - an ingredient name
- * @param {String} id - an id of an H2 to look for in the hast tree
- * @returns {Function} a function
- */
-function requireTopLevelHeading(id) {
-  return (tree, file, context) => {
-    const heading = select(`h2#${id}`, tree);
-    if (heading === null) {
-      utils.logMissingIngredient(file, context);
-    }
-  };
-}
-
-module.exports = ingredientHandlers;
