+++ scripts/scraper-ng/rules/html-require-recipe-ingredients/ingredient-handlers/data-browser-compatibility.js
@@ -1,0 +1,36 @@
+const { select } = require("hast-util-select");
+const visit = require("unist-util-visit");
+
+const utils = require("./utils.js");
+
+function handleDataBrowserCompatibility(tree, file, context) {
+  const id = "Browser_compatibility";
+  const body = select(`body`, tree);
+  const heading = select(`h2#${id}`, body);
+
+  if (heading === null) {
+    const message = file.message(
+      `Expected h2#${id}`,
+      body,
+      `${context.source}:${context.recipeName}/${context.ingredient}/expected-heading`
+    );
+    message.fatal = true;
+    utils.logMissingIngredient(file, context);
+    return;
+  }
+
+  let macroCount = 0;
+  visit(
+    utils.sliceSection(heading, body),
+    node => utils.isMacro(node, "Compat"),
+    () => {
+      macroCount += 1;
+    }
+  );
+
+  if (macroCount !== 1) {
+    utils.logMissingIngredient(file, context);
+  }
+}
+
+module.exports = handleDataBrowserCompatibility;
